requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.2
base58==2.1.0
bitarray==1.2.2
blinker==1.4
CacheControl==0.12.10
cachelib==0.1
cachetools==4.2.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-tester==0.5.0b4
eth-typing==2.2.2
eth-utils==1.10.0
Faker==9.5.2
firebase-admin==5.1.0
Flask==1.1.2
Flask-And-Redis==1.0.0
Flask-JWT-Extended==4.3.1
Flask-Limiter==1.4
flask-marshmallow==0.14.0
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
gevent==21.1.2
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.42.0
grpcio-status==1.42.0
hexbytes==0.2.2
httplib2==0.20.2
pyhumps==3.5.3
idna==2.10
ipfshttpclient==0.8.0a2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
limits==1.5.1
lru-dict==1.1.7
MarkupSafe==1.1.1
marshmallow==3.11.1
msgpack==0.6.2
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
proto-plus==1.19.8
protobuf==3.19.0
py3rijndael==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.11.0
pydash==5.0.0
PyJWT==2.1.0
pymongo==3.9.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.16.0
python-slugify==4.0.0
redis==3.0.0
redis-py-cluster==2.1.2
requests==2.25.1
rlp==2.0.1
rsa==4.7.2
secure-smtplib==0.1.1
semantic-version==2.8.5
sentry-sdk==1.0.0
six==1.15.0
text-unidecode==1.3
toolz==0.11.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.4
uWSGI==2.0.19.1
varint==1.0.2
web3==5.24.0
websockets==9.1
Werkzeug==1.0.1
yarl==1.7.0
zope.event==4.5.0
zope.interface==5.4.0
